Refactored challenge agent logic.
