+++ pylot/simulation/challenge/ERDOSAgent.py
@@ -1,35 +1,23 @@
-import logging
-from collections import deque
-
 from absl import flags
-
-import carla
 
 import erdos
 
-from leaderboard.autoagents.autonomous_agent import AutonomousAgent, \
-    Track
-
-import numpy as np
+from leaderboard.autoagents.autonomous_agent import Track
 
 import pylot.flags
 import pylot.component_creator  # noqa: I100
 import pylot.operator_creator
 import pylot.perception.messages
-import pylot.simulation.utils
 import pylot.utils
 from pylot.drivers.sensor_setup import LidarSetup, RGBCameraSetup
-from pylot.localization.messages import GNSSMessage, IMUMessage
 from pylot.perception.camera_frame import CameraFrame
-from pylot.perception.messages import ObstaclesMessage, TrafficLightsMessage
-from pylot.perception.point_cloud import PointCloud
-from pylot.planning.messages import WaypointsMessage
-from pylot.planning.waypoints import Waypoints
+from pylot.simulation.challenge.ERDOSBaseAgent import ERDOSBaseAgent, \
+    process_visualization_events, read_control_command
 
 FLAGS = flags.FLAGS
 
 flags.DEFINE_bool('carla_localization', False,
-                  'True to use perfect localization')
+                  'Set to True to receive ego-vehicle locations from CARLA')
 
 
 # Certain visualizations are not supported when running in challenge mode.
@@ -47,200 +35,71 @@ flags.register_multi_flags_validator(
     unsupported_visualizations_validator,
     message='Trying to visualize unsupported_visualization')
 
+# Location of the center camera relative to the ego-vehicle.
 CENTER_CAMERA_LOCATION = pylot.utils.Location(0.0, 0.0, 2.0)
 CENTER_CAMERA_NAME = 'center_camera'
+# Location of the camera used for lane detection.
 LANE_CAMERA_LOCATION = pylot.utils.Location(1.3, 0.0, 1.8)
 LANE_CAMERA_NAME = 'lane_camera'
 TL_CAMERA_NAME = 'traffic_lights_camera'
-LEFT_CAMERA_NAME = 'left_camera'
-RIGHT_CAMERA_NAME = 'right_camera'
 
 
 def get_entry_point():
     return 'ERDOSAgent'
 
 
-def using_perfect_component():
-    """Returns True if the agent uses any perfect component."""
-    return (FLAGS.carla_obstacle_detection
-            or FLAGS.carla_traffic_light_detection
-            or FLAGS.perfect_obstacle_tracking or FLAGS.carla_localization)
-
-
-def using_lidar():
-    """Returns True if Lidar is required for the setup."""
-    return not (FLAGS.carla_obstacle_detection
-                and FLAGS.carla_traffic_light_detection)
-
-
-class ERDOSAgent(AutonomousAgent):
-    """Agent class that interacts with the CARLA challenge scenario runner.
+class ERDOSAgent(ERDOSBaseAgent):
+    """Agent class that interacts with the CARLA challenge leaderboard.
 
     Attributes:
-        track: Track the agent is running in.
         _camera_setups: Mapping between camera names and
             :py:class:`~pylot.drivers.sensor_setup.CameraSetup`.
-        _lidar_transform (:py:class:`~pylot.utils.Transform`): Transform of
-            the Lidar relative to the ego vehicle.
-        _waypoints (list(:py:class:`~pylot.utils.Transform`)): List of
-            waypoints the agent receives from the challenge planner.
+        _lidar_setup (:py:class:`~pylot.drivers.sensor_setup.LidarSetup`):
+            Setup of the Lidar sensor.
     """
     def setup(self, path_to_conf_file):
-        """Setup phase. Invoked by the scenario runner."""
-        # Disable Tensorflow logging.
-        pylot.utils.set_tf_loglevel(logging.ERROR)
-        # Parse the flag file.
-        flags.FLAGS([__file__, '--flagfile={}'.format(path_to_conf_file)])
-        self._logger = erdos.utils.setup_logging('erdos_agent',
-                                                 FLAGS.log_file_name)
-        enable_logging()
-        self.track = get_track()
-        self._town_name = None
+        super(ERDOSAgent, self).setup(path_to_conf_file)
         self._camera_setups = create_camera_setups()
-        # Set the lidar in the same position as the center camera.
-        self._lidar_transform = pylot.utils.Transform(CENTER_CAMERA_LOCATION,
-                                                      pylot.utils.Rotation())
-        self._lidar_setup = LidarSetup('lidar',
-                                       'sensor.lidar.ray_cast',
-                                       self._lidar_transform,
-                                       range=8500,
-                                       legacy=False)
-        self._last_point_cloud = None
-        self._last_yaw = 0
-        # Stores the waypoints we get from the challenge planner.
-        self._waypoints = None
-        # Stores the open drive string we get when we run in track 3.
-        self._open_drive_data = None
-        (camera_streams, pose_stream, route_stream, global_trajectory_stream,
-         open_drive_stream, point_cloud_stream, imu_stream, gnss_stream,
-         control_stream, control_display_stream, ground_obstacles_stream,
-         ground_traffic_lights_stream, vehicle_id_stream,
-         streams_to_send_top_on) = create_data_flow()
-        self._camera_streams = camera_streams
-        self._pose_stream = pose_stream
-        self._route_stream = route_stream
-        self._sent_initial_pose = False
-        self._global_trajectory_stream = global_trajectory_stream
-        self._open_drive_stream = open_drive_stream
+        self._lidar_setup = create_lidar_setup()
+
         self._sent_open_drive = False
-        self._point_cloud_stream = point_cloud_stream
-        self._imu_stream = imu_stream
-        self._gnss_stream = gnss_stream
-        self._control_stream = control_stream
-        self._control_display_stream = control_display_stream
-        self._ground_obstacles_stream = ground_obstacles_stream
-        self._ground_traffic_lights_stream = ground_traffic_lights_stream
-        self._vehicle_id_stream = vehicle_id_stream
-        self._ego_vehicle = None
-        self._sent_vehicle_id = False
+
+        # Create the dataflow of AV components. Change the method
+        # to add your operators.
+        (self._camera_streams, self._pose_stream, self._route_stream,
+         self._global_trajectory_stream, self._open_drive_stream,
+         self._point_cloud_stream, self._imu_stream, self._gnss_stream,
+         self._control_stream, self._control_display_stream,
+         self._perfect_obstacles_stream, self._perfect_traffic_lights_stream,
+         self._vehicle_id_stream, streams_to_send_top_on) = create_data_flow()
         # Execute the dataflow.
         self._node_handle = erdos.run_async()
+        # Close the streams that are not used (i.e., send top watermark).
         for stream in streams_to_send_top_on:
             stream.send(erdos.WatermarkMessage(erdos.Timestamp(is_top=True)))
-        if using_perfect_component():
-            # The agent is using a perfect component. It must directly connect
-            # to the simulator to send perfect data to the data-flow.
-            _, self._world = pylot.simulation.utils.get_world(
-                FLAGS.carla_host, FLAGS.carla_port, FLAGS.carla_timeout)
 
     def destroy(self):
         """Clean-up the agent. Invoked between different runs."""
-        self._logger.info('ERDOSAgent destroy method invoked')
+        self.logger.info('ERDOSAgent destroy method invoked')
+        # Stop the ERDOS node.
         self._node_handle.shutdown()
+        # Reset the ERDOS dataflow graph.
         erdos.reset()
-
-    def sensors(self):
-        """
-        Defines the sensor suite required by the agent.
-        """
-        opendrive_map_sensors = []
-        if self.track == Track.MAP:
-            opendrive_map_sensors = [{
-                'type': 'sensor.opendrive_map',
-                'reading_frequency': 20,
-                'id': 'opendrive'
-            }]
-
-        gnss_sensors = [{
-            'type': 'sensor.other.gnss',
-            'x': 0.0,
-            'y': 0.0,
-            'z': 0.0,
-            'id': 'gnss'
-        }]
-
-        imu_sensors = [{
-            'type': 'sensor.other.imu',
-            'x': 0.0,
-            'y': 0.0,
-            'z': 0.0,
-            'roll': 0.0,
-            'pitch': 0.0,
-            'yaw': 0.0,
-            'id': 'imu'
-        }]
-
-        speed_sensors = [{
-            'type': 'sensor.speedometer',
-            'reading_frequency': 20,
-            'id': 'speed'
-        }]
-
-        if using_lidar():
-            lidar_sensors = [{
-                'type': 'sensor.lidar.ray_cast',
-                'x': self._lidar_transform.location.x,
-                'y': self._lidar_transform.location.y,
-                'z': self._lidar_transform.location.z,
-                'roll': self._lidar_transform.rotation.roll,
-                'pitch': self._lidar_transform.rotation.pitch,
-                'yaw': self._lidar_transform.rotation.yaw,
-                'id': 'LIDAR'
-            }]
-        else:
-            lidar_sensors = []
-
-        camera_sensors = []
-        for cs in self._camera_setups.values():
-            camera_sensor = {
-                'type': cs.camera_type,
-                'x': cs.transform.location.x,
-                'y': cs.transform.location.y,
-                'z': cs.transform.location.z,
-                'roll': cs.transform.rotation.roll,
-                'pitch': cs.transform.rotation.pitch,
-                'yaw': cs.transform.rotation.yaw,
-                'width': cs.width,
-                'height': cs.height,
-                'fov': cs.fov,
-                'id': cs.name
-            }
-            camera_sensors.append(camera_sensor)
-
-        return (gnss_sensors + speed_sensors + imu_sensors +
-                opendrive_map_sensors + camera_sensors + lidar_sensors)
 
     def run_step(self, input_data, timestamp):
         game_time = int(timestamp * 1000)
-        self._logger.debug("Current game time {}".format(game_time))
         erdos_timestamp = erdos.Timestamp(coordinates=[game_time])
 
-        if not self._sent_open_drive and self.track != Track.MAP:
-            # We do not have access to the open drive map. Send top watermark.
-            self._sent_open_drive = True
-            self._open_drive_stream.send(
-                erdos.WatermarkMessage(erdos.Timestamp(is_top=True)))
-        # Send the waypoints.
-        self.send_waypoints_msg(erdos_timestamp)
-
+        # Parse the sensor data the agent receives from the leaderboard.
         speed_data = None
         imu_data = None
         gnss_data = None
-        carla_pc = None
         for key, val in input_data.items():
-            # val is a tuple of (data timestamp, data).
-            # print("Sensor {} at {}".format(key, val[0]))
+            # val is a tuple of (timestamp, data).
             if key in self._camera_streams:
+                # The data is for one of the cameras. Transform it to a Pylot
+                # CameraFrame, and send it on the camera's corresponding
+                # stream.
                 self._camera_streams[key].send(
                     pylot.perception.messages.FrameMessage(
                         erdos_timestamp,
@@ -255,296 +114,176 @@ class ERDOSAgent(AutonomousAgent):
             elif key == 'gnss':
                 gnss_data = val[1]
             elif key == 'opendrive':
-                self.send_opendrive_map_msg(val[1], erdos_timestamp)
+                if not self._open_drive_stream.is_closed():
+                    # The data is only sent once because it does not change
+                    # throught the duration of a scenario.
+                    self._open_drive_stream.send(
+                        erdos.Message(erdos_timestamp, val[1]['opendrive']))
+                    # Inform the operators that read the open drive stream that
+                    # they will not receive any other messages on this stream.
+                    self._open_drive_stream.send(
+                        erdos.WatermarkMessage(erdos.Timestamp(is_top=True)))
             elif key == 'LIDAR':
+                # Send the LiDAR point cloud.
                 carla_pc = val[1]
+                self.send_lidar_msg(self._point_cloud_stream, carla_pc,
+                                    erdos_timestamp, self._lidar_setup)
             else:
-                self._logger.warning("Sensor {} not used".format(key))
+                self.logger.warning("Sensor {} not used".format(key))
 
-        self.send_vehicle_id_msg()
-        self.send_ground_data(erdos_timestamp)
+        # Send the route the vehicle must follow.
+        self.send_global_trajectory_msg(self._global_trajectory_stream,
+                                        erdos_timestamp)
 
+        # The following two methods are only relevant when the agent
+        # is using perfect perception.
+        self.send_vehicle_id_msg(self._vehicle_id_stream)
+        self.send_perfect_detections(self._perfect_obstacles_stream,
+                                     self._perfect_traffic_lights_stream,
+                                     erdos_timestamp, CENTER_CAMERA_LOCATION)
+
+        # Send localization data.
+        self.send_localization(erdos_timestamp, imu_data, gnss_data,
+                               speed_data)
+
+        # Ensure that the open drive stream is closed when the agent
+        # is not running on the MAP track.
+        if not self._open_drive_stream.is_closed() and self.track != Track.MAP:
+            # We do not have access to the open drive map. Send top watermark.
+            self._open_drive_stream.send(
+                erdos.WatermarkMessage(erdos.Timestamp(is_top=True)))
+
+        process_visualization_events(self._control_display_stream)
+
+        # Return the control command received on the control stream.
+        return read_control_command(self._control_stream)
+
+    def send_localization(self, timestamp, imu_data, gnss_data, speed_data):
         if FLAGS.localization:
-            self.send_imu_msg(imu_data, erdos_timestamp)
-            self.send_gnss_msg(gnss_data, erdos_timestamp)
+            # The agent uses our localization. We need to send data on the
+            # IMU and GNSS streams, and the initial position of the ego-vehicle
+            # on the route stream.
+            self.send_imu_msg(self._imu_stream, imu_data, timestamp)
+            self.send_gnss_msg(self._gnss_stream, gnss_data, timestamp)
+            # Naively compute the pose of the ego-vehicle, and send it on
+            # the route stream. Pylot's localization operator will refine this
+            # pose using the GNSS and IMU data.
             pose = self.compute_pose(speed_data, imu_data, gnss_data,
-                                     erdos_timestamp)
-            self._route_stream.send(erdos.Message(erdos_timestamp, pose))
-            self._route_stream.send(erdos.WatermarkMessage(erdos_timestamp))
+                                     timestamp)
+            self._route_stream.send(erdos.Message(timestamp, pose))
+            self._route_stream.send(erdos.WatermarkMessage(timestamp))
         elif FLAGS.carla_localization:
-            self.send_perfect_pose_msg(erdos_timestamp)
+            self.send_perfect_pose_msg(self._pose_stream, timestamp)
         else:
+            # In this configuration, the agent is not using a localization
+            # operator. It is driving using the noisy localization it receives
+            # from the leaderboard.
             pose = self.compute_pose(speed_data, imu_data, gnss_data,
-                                     erdos_timestamp)
-            self._pose_stream.send(erdos.Message(erdos_timestamp, pose))
-            self._pose_stream.send(erdos.WatermarkMessage(erdos_timestamp))
-
-        if using_lidar():
-            self.send_lidar_msg(carla_pc, self._lidar_transform,
-                                erdos_timestamp)
-
-        if pylot.flags.must_visualize():
-            import pygame
-            from pygame.locals import K_n
-            events = pygame.event.get()
-            for event in events:
-                if event.type == pygame.KEYUP:
-                    if event.key == K_n:
-                        self._control_display_stream.send(
-                            erdos.Message(erdos.Timestamp(coordinates=[0]),
-                                          event.key))
-
-        # Wait until the control is set.
-        while True:
-            control_msg = self._control_stream.read()
-            if not isinstance(control_msg, erdos.WatermarkMessage):
-                output_control = carla.VehicleControl()
-                output_control.throttle = control_msg.throttle
-                output_control.brake = control_msg.brake
-                output_control.steer = control_msg.steer
-                output_control.reverse = control_msg.reverse
-                output_control.hand_brake = control_msg.hand_brake
-                output_control.manual_gear_shift = False
-                return output_control
-
-    def send_gnss_msg(self, gnss_data, timestamp):
-        latitude = gnss_data[0]
-        longitude = gnss_data[1]
-        altitude = gnss_data[2]
-        location = pylot.utils.Location.from_gps(latitude, longitude, altitude)
-        transform = pylot.utils.Transform(location, pylot.utils.Rotation())
-        msg = GNSSMessage(timestamp, transform, altitude, latitude, longitude)
-        self._gnss_stream.send(msg)
-        self._gnss_stream.send(erdos.WatermarkMessage(timestamp))
-
-    def send_imu_msg(self, imu_data, timestamp):
-        accelerometer = pylot.utils.Vector3D(imu_data[0], imu_data[1],
-                                             imu_data[2])
-        gyroscope = pylot.utils.Vector3D(imu_data[3], imu_data[4], imu_data[5])
-        compass = imu_data[6]
-        msg = IMUMessage(timestamp, None, accelerometer, gyroscope, compass)
-        self._imu_stream.send(msg)
-        self._imu_stream.send(erdos.WatermarkMessage(timestamp))
-
-    def send_opendrive_map_msg(self, data, timestamp):
-        # Sending once opendrive data
-        if self._open_drive_data is None:
-            self._open_drive_data = data['opendrive']
-            self._open_drive_stream.send(
-                erdos.Message(timestamp, self._open_drive_data))
-            self._open_drive_stream.send(
-                erdos.WatermarkMessage(erdos.Timestamp(is_top=True)))
-        else:
-            self._logger.warning(
-                'Agent did not sent open drive data for {}'.format(timestamp))
-
-    def compute_pose(self, speed_data, imu_data, gnss_data, timestamp):
-        forward_speed = speed_data['speed']
-        # vehicle_transform = pylot.utils.Transform.from_carla_transform(
-        #     speed_data['transform'])
-        latitude = gnss_data[0]
-        longitude = gnss_data[1]
-        altitude = gnss_data[2]
-        location = pylot.utils.Location.from_gps(latitude, longitude, altitude)
-        if np.isnan(imu_data[6]):
-            yaw = self._last_yaw
-        else:
-            compass = np.degrees(imu_data[6])
-            if compass < 270:
-                yaw = compass - 90
-            else:
-                yaw = compass - 450
-            self._last_yaw = yaw
-        vehicle_transform = pylot.utils.Transform(
-            location, pylot.utils.Rotation(yaw=yaw))
-        velocity_vector = pylot.utils.Vector3D(forward_speed * np.cos(yaw),
-                                               forward_speed * np.sin(yaw), 0)
-        current_pose = pylot.utils.Pose(vehicle_transform, forward_speed,
-                                        velocity_vector,
-                                        timestamp.coordinates[0])
-        self._logger.debug("@{}: Predicted pose: {}".format(
-            timestamp, current_pose))
-        return current_pose
-
-    def send_lidar_msg(self,
-                       carla_pc,
-                       transform,
-                       timestamp,
-                       ego_transform=None):
-        # Remove the intensity component of the point cloud.
-        carla_pc = carla_pc[:, :3]
-        point_cloud = PointCloud(carla_pc, self._lidar_setup)
-        if self._last_point_cloud is not None:
-            # TODO(ionel): Should offset the last point cloud wrt to the
-            # current location.
-            # self._last_point_cloud.global_points = \
-            #     ego_transform.transform_points(
-            #         self._last_point_cloud.global_points)
-            # self._last_point_cloud.points = \
-            #     self._last_point_cloud._to_camera_coordinates(
-            #         self._last_point_cloud.global_points)
-            point_cloud.merge(self._last_point_cloud)
-        self._point_cloud_stream.send(
-            pylot.perception.messages.PointCloudMessage(
-                timestamp, point_cloud))
-        self._point_cloud_stream.send(erdos.WatermarkMessage(timestamp))
-        # global_pc = ego_transform.inverse_transform_points(carla_pc)
-        self._last_point_cloud = PointCloud(carla_pc, self._lidar_setup)
-
-    def send_waypoints_msg(self, timestamp):
-        # Send once the global waypoints.
-        if self._waypoints is None:
-            # Gets global waypoints from the agent.
-            waypoints = deque([])
-            road_options = deque([])
-            for (transform, road_option) in self._global_plan_world_coord:
-                waypoints.append(
-                    pylot.utils.Transform.from_carla_transform(transform))
-                road_options.append(pylot.utils.RoadOption(road_option.value))
-            self._waypoints = Waypoints(waypoints, road_options=road_options)
-            self._global_trajectory_stream.send(
-                WaypointsMessage(timestamp, self._waypoints))
-            self._global_trajectory_stream.send(
-                erdos.WatermarkMessage(erdos.Timestamp(is_top=True)))
-
-    def send_vehicle_id_msg(self):
-        if ((FLAGS.carla_localization or FLAGS.perfect_obstacle_tracking)
-                and not self._sent_vehicle_id):
-            actor_list = self._world.get_actors()
-            vec_actors = actor_list.filter('vehicle.*')
-            for actor in vec_actors:
-                if ('role_name' in actor.attributes
-                        and actor.attributes['role_name'] == 'hero'):
-                    self._ego_vehicle = actor
-                    break
-            if self._ego_vehicle:
-                self._sent_vehicle_id = True
-                self._vehicle_id_stream.send(
-                    erdos.Message(erdos.Timestamp(coordinates=[0]),
-                                  self._ego_vehicle.id))
-                self._vehicle_id_stream.send(
-                    erdos.WatermarkMessage(erdos.Timestamp(is_top=True)))
-
-    def send_perfect_pose_msg(self, timestamp):
-        vec_transform = pylot.utils.Transform.from_carla_transform(
-            self._ego_vehicle.get_transform())
-        velocity_vector = pylot.utils.Vector3D.from_carla_vector(
-            self._ego_vehicle.get_velocity())
-        forward_speed = velocity_vector.magnitude()
-        pose = pylot.utils.Pose(vec_transform, forward_speed, velocity_vector,
-                                timestamp.coordinates[0])
-        self._pose_stream.send(erdos.Message(timestamp, pose))
-        self._pose_stream.send(erdos.WatermarkMessage(timestamp))
-
-    def send_ground_data(self, timestamp):
-        if not (FLAGS.carla_obstacle_detection
-                or FLAGS.carla_traffic_light_detection):
-            return
-        actor_list = self._world.get_actors()
-        (vehicles, people, traffic_lights, _,
-         _) = pylot.simulation.utils.extract_data_in_pylot_format(actor_list)
-        if FLAGS.carla_obstacle_detection:
-            self._ground_obstacles_stream.send(
-                ObstaclesMessage(timestamp, vehicles + people))
-            self._ground_obstacles_stream.send(
-                erdos.WatermarkMessage(timestamp))
-        if FLAGS.carla_traffic_light_detection:
-            vec_transform = pylot.utils.Transform.from_carla_transform(
-                self._ego_vehicle.get_transform())
-            tl_camera_transform = pylot.utils.Transform(
-                CENTER_CAMERA_LOCATION, pylot.utils.Rotation())
-            visible_tls = []
-            if self._town_name is None:
-                self._town_name = self._world.get_map().name
-            for tl in traffic_lights:
-                if tl.is_traffic_light_visible(
-                        vec_transform * tl_camera_transform,
-                        self._town_name,
-                        distance_threshold=FLAGS.
-                        static_obstacle_distance_threshold):
-                    if self._town_name not in ['Town01', 'Town02']:
-                        delta_y = -5
-                        if self._town_name == 'Town04':
-                            delta_y = -2
-                        # Move the traffic light location to the road.
-                        tl.transform = tl.transform * pylot.utils.Transform(
-                            pylot.utils.Location(delta_y, 0, 5),
-                            pylot.utils.Rotation())
-                    visible_tls.append(tl)
-            self._ground_traffic_lights_stream.send(
-                TrafficLightsMessage(timestamp, visible_tls))
-            self._ground_traffic_lights_stream.send(
-                erdos.WatermarkMessage(timestamp))
-
-
-def get_track():
-    track = None
-    if FLAGS.execution_mode == 'challenge-sensors':
-        track = Track.SENSORS
-    elif FLAGS.execution_mode == 'challenge-map':
-        track = Track.MAP
-    else:
-        raise ValueError('Unexpected track {}'.format(FLAGS.execution_mode))
-    return track
+                                     timestamp)
+            self._pose_stream.send(erdos.Message(timestamp, pose))
+            self._pose_stream.send(erdos.WatermarkMessage(timestamp))
 
 
 def create_data_flow():
-    """ Create the challenge data-flow graph."""
+    """Creates a dataflow graph of operators.
+
+    This is the place to add other operators (e.g., other detectors,
+    behavior planning).
+    """
     streams_to_send_top_on = []
-    time_to_decision_loop_stream = erdos.LoopStream()
     camera_setups = create_camera_setups()
+    # Creates a dataflow stream for each camera.
     camera_streams = {}
     for name in camera_setups:
         camera_streams[name] = erdos.IngestStream()
+    # Creates a stream on which the agent sends the high-level route the
+    # agent must follow in the challenge.
     global_trajectory_stream = erdos.IngestStream()
+    # Creates a stream on which the agent sends the open drive stream it
+    # receives when it executes in the MAP track.
     open_drive_stream = erdos.IngestStream()
+    # Creates a stream on which the agent sends point cloud messages it
+    # receives from the LiDAR sensor.
     point_cloud_stream = erdos.IngestStream()
     imu_stream = erdos.IngestStream()
     gnss_stream = erdos.IngestStream()
     route_stream = erdos.IngestStream()
+    time_to_decision_loop_stream = erdos.LoopStream()
 
     if FLAGS.localization:
+        # Pylot localization is enabled. Add the localization operator to
+        # the dataflow. The operator receives GNSS and IMU messages, and
+        # uses an Extended Kalman Filter to compute poses.
         pose_stream = pylot.operator_creator.add_localization(
             imu_stream, gnss_stream, route_stream)
     else:
+        # The agent either directly forwards the poses it receives from
+        # the challenge, which are noisy, or the perfect poses if the
+        # --carla_localization flag is set.
         pose_stream = erdos.IngestStream()
 
-    ground_obstacles_stream = erdos.IngestStream()
-    ground_traffic_lights_stream = erdos.IngestStream()
-    vehicle_id_stream = erdos.IngestStream()
-    if not (FLAGS.perfect_obstacle_tracking or FLAGS.carla_localization):
-        streams_to_send_top_on.append(vehicle_id_stream)
-
+    # Stream on which the obstacles are sent when the agent is using perfect
+    # detection.
+    perfect_obstacles_stream = erdos.IngestStream()
     if FLAGS.carla_obstacle_detection:
-        obstacles_stream = ground_obstacles_stream
+        # Execute with perfect perception. In this configuration, the agent
+        # directly gets the location of the other agents from CARLA. This
+        # configuration is meant for testing and debugging.
+        obstacles_stream = perfect_obstacles_stream
     elif any('efficientdet' in model
              for model in FLAGS.obstacle_detection_model_names):
+        # Add an operator that runs EfficientDet to detect obstacles.
         obstacles_stream = pylot.operator_creator.\
             add_efficientdet_obstacle_detection(
                 camera_streams[CENTER_CAMERA_NAME],
                 time_to_decision_loop_stream)[0]
-        streams_to_send_top_on.append(ground_obstacles_stream)
+        streams_to_send_top_on.append(perfect_obstacles_stream)
     else:
+        # Add an operator that uses one of the Tensorflow model zoo
+        # detection models. By default, we use FasterRCNN.
         obstacles_stream = pylot.operator_creator.add_obstacle_detection(
             camera_streams[CENTER_CAMERA_NAME],
             time_to_decision_loop_stream)[0]
-        streams_to_send_top_on.append(ground_obstacles_stream)
+        streams_to_send_top_on.append(perfect_obstacles_stream)
 
+    # Stream on which the traffic lights are sent when the agent is
+    # using perfect traffic light detection.
+    perfect_traffic_lights_stream = erdos.IngestStream()
     if FLAGS.carla_traffic_light_detection:
-        traffic_lights_stream = ground_traffic_lights_stream
+        # In this debug configuration, the agent is using perfectly located
+        # traffic lights it receives directly from CARLA. Therefore, there's
+        # no need to a traffic light detector.
+        traffic_lights_stream = perfect_traffic_lights_stream
         camera_streams[TL_CAMERA_NAME] = erdos.IngestStream()
         streams_to_send_top_on.append(camera_streams[TL_CAMERA_NAME])
     else:
+        # Adds a traffic light detector operator, which uses the camera with
+        # the small fov.
         traffic_lights_stream = \
             pylot.operator_creator.add_traffic_light_detector(
                 camera_streams[TL_CAMERA_NAME])
         # Adds an operator that finds the world location of the traffic lights.
+        # The operator synchronizes LiDAR point cloud readings with camera
+        # frames, and uses them to compute the depth to traffic light bounding
+        # boxes.
         traffic_lights_stream = \
             pylot.operator_creator.add_obstacle_location_finder(
                 traffic_lights_stream, point_cloud_stream, pose_stream,
                 camera_setups[TL_CAMERA_NAME])
-        streams_to_send_top_on.append(ground_traffic_lights_stream)
+        # We do not send perfectly located traffic lights in this
+        # configuration. Therefore, ensure that the stream is "closed"
+        # (i.e., send a top watermark)
+        streams_to_send_top_on.append(perfect_traffic_lights_stream)
 
+    vehicle_id_stream = erdos.IngestStream()
+    if not (FLAGS.perfect_obstacle_tracking or FLAGS.carla_localization):
+        # The vehicle_id_stream is only used when perfect CARLA localization
+        # or perfect obstacle tracking are enabled.
+        streams_to_send_top_on.append(vehicle_id_stream)
+
+    # Adds an operator for tracking detected agents. The operator uses the
+    # frames from the center camera, and the bounding boxes found by the
+    # obstacle detector operator.
     obstacles_tracking_stream = pylot.component_creator.add_obstacle_tracking(
         camera_streams[CENTER_CAMERA_NAME],
         camera_setups[CENTER_CAMERA_NAME],
@@ -552,25 +291,41 @@ def create_data_flow():
         depth_stream=point_cloud_stream,
         vehicle_id_stream=vehicle_id_stream,
         pose_stream=pose_stream,
-        ground_obstacles_stream=ground_obstacles_stream,
+        ground_obstacles_stream=perfect_obstacles_stream,
         time_to_decision_stream=time_to_decision_loop_stream)
 
     if FLAGS.execution_mode == 'challenge-sensors':
+        # The agent is running is sensors-only track. Therefore, we need
+        # to add a lane detector because the agent does not have access to
+        # the OpenDrive map.
         lanes_stream = pylot.operator_creator.add_lanenet_detection(
             camera_streams[LANE_CAMERA_NAME])
     else:
+        # The lanes stream is not used when running in the Map track.
+        # We add the stream to the list of streams that are not used, and
+        # must be manually "closed" (i.e., send a top watermark).
         lanes_stream = erdos.IngestStream()
         streams_to_send_top_on.append(lanes_stream)
 
+    # The agent uses a linear predictor to compute future trajectories
+    # of the other agents.
     prediction_stream = pylot.operator_creator.add_linear_prediction(
         obstacles_tracking_stream)
 
+    # Adds a planner to the agent. The planner receives the pose of
+    # the ego-vehicle, detected traffic lights, predictions for other
+    # agents, the route the agent must follow, and the open drive data if
+    # the agent is executing in the Map track, or detected lanes if it is
+    # executing in the Sensors-only track.
     waypoints_stream = pylot.component_creator.add_planning(
         None, pose_stream, prediction_stream, traffic_lights_stream,
         lanes_stream, open_drive_stream, global_trajectory_stream,
         time_to_decision_loop_stream)
 
     if pylot.flags.must_visualize():
+        # Adds a visualization dataflow operator if any of the
+        # --visualize_* is enabled. The operator creates a pygame window
+        # in which the different sensors, detections can be visualized.
         control_display_stream, ingest_streams = \
             pylot.operator_creator.add_visualizer(
                 pose_stream=pose_stream,
@@ -587,10 +342,18 @@ def create_data_flow():
     else:
         control_display_stream = None
 
+    # Adds a PID controller which tries to follow the waypoints computed
+    # by the planner.
     control_stream = pylot.operator_creator.add_pid_control(
         pose_stream, waypoints_stream)
+    # The PID planner returns a stream of commands (i.e., throttle, steer)
+    # from which the agent can read the command it must return to the
+    # challenge.
     extract_control_stream = erdos.ExtractStream(control_stream)
 
+    # Operator that computes how much time each component gets to execute.
+    # This is needed in Pylot, but can be ignored when running in challenge
+    # mode.
     time_to_decision_stream = pylot.operator_creator.add_time_to_decision(
         pose_stream, obstacles_stream)
     time_to_decision_loop_stream.set(time_to_decision_stream)
@@ -598,13 +361,22 @@ def create_data_flow():
     return (camera_streams, pose_stream, route_stream,
             global_trajectory_stream, open_drive_stream, point_cloud_stream,
             imu_stream, gnss_stream, extract_control_stream,
-            control_display_stream, ground_obstacles_stream,
-            ground_traffic_lights_stream, vehicle_id_stream,
+            control_display_stream, perfect_obstacles_stream,
+            perfect_traffic_lights_stream, vehicle_id_stream,
             streams_to_send_top_on)
 
 
 def create_camera_setups():
+    """Creates RGBCameraSetups for the cameras deployed on the car.
+
+    Note: Change this method if your agent requires more cameras, or if
+    you want to change camera properties.
+
+    It returns a dict from camera name to
+    :py:class:`~pylot.drivers.sensor_setup.RGBCameraSetup`.
+    """
     camera_setups = {}
+    # Add a center front camera.
     transform = pylot.utils.Transform(CENTER_CAMERA_LOCATION,
                                       pylot.utils.Rotation())
     center_camera_setup = RGBCameraSetup(CENTER_CAMERA_NAME,
@@ -612,42 +384,40 @@ def create_camera_setups():
                                          FLAGS.camera_image_height, transform,
                                          90)
     camera_setups[CENTER_CAMERA_NAME] = center_camera_setup
+
     if not FLAGS.carla_traffic_light_detection:
+        # Add a camera with a narrow field of view. The traffic light
+        # camera is added in the same position as the center camera.
+        # We use this camera for traffic light detection.
         tl_camera_setup = RGBCameraSetup(TL_CAMERA_NAME,
                                          FLAGS.camera_image_width,
                                          FLAGS.camera_image_height, transform,
                                          45)
         camera_setups[TL_CAMERA_NAME] = tl_camera_setup
+
     if FLAGS.execution_mode == 'challenge-sensors':
-        # Add camera for lane detection.
+        # The agent in executed in the sensors-only track.
+        # We add camera, which we use for lane detection because we do not
+        # have access to the OpenDrive map in this track.
         lane_transform = pylot.utils.Transform(LANE_CAMERA_LOCATION,
                                                pylot.utils.Rotation(pitch=-15))
         lane_camera_setup = RGBCameraSetup(LANE_CAMERA_NAME, 1280, 720,
                                            lane_transform, 90)
         camera_setups[LANE_CAMERA_NAME] = lane_camera_setup
-
-    # left_transform = pylot.utils.Transform(CENTER_CAMERA_LOCATION,
-    #                                        pylot.utils.Rotation(yaw=-45))
-    # left_camera_setup = RGBCameraSetup(LEFT_CAMERA_NAME,
-    #                                    FLAGS.camera_image_width,
-    #                                    FLAGS.camera_image_height,
-    #                                    left_transform, 90)
-    # camera_setups[LEFT_CAMERA_NAME] = left_camera_setup
-    # right_transform = pylot.utils.Transform(CENTER_CAMERA_LOCATION,
-    #                                         pylot.utils.Rotation(yaw=45))
-    # right_camera_setup = RGBCameraSetup(RIGHT_CAMERA_NAME,
-    #                                     FLAGS.camera_image_width,
-    #                                     FLAGS.camera_image_height,
-    #                                     right_transform, 90)
-    # camera_setups[RIGHT_CAMERA_NAME] = right_camera_setup
     return camera_setups
 
 
-def enable_logging():
-    """Overwrites logging config so that loggers can control verbosity.
-
-    This method is required because the challenge evaluator overwrites
-    verbosity, which causes Pylot log messages to be discarded.
-    """
-    import logging
-    logging.root.setLevel(logging.NOTSET)
+def create_lidar_setup():
+    """Creates a setup for the LiDAR deployed on the car."""
+    # Set the lidar in the same position as the center camera.
+    # Pylot uses the LiDAR point clouds to compute the distance to
+    # the obstacles detected using the camera frames.
+    lidar_transform = pylot.utils.Transform(CENTER_CAMERA_LOCATION,
+                                            pylot.utils.Rotation())
+    lidar_setup = LidarSetup(
+        'lidar',
+        'sensor.lidar.ray_cast',
+        lidar_transform,
+        range=8500,  # Range is only used for visualization.
+        legacy=False)
+    return lidar_setup
